+++ lib/onkcop.rb
@@ -1,6 +1,0 @@
-require "onkcop/cli"
-require "onkcop/version"
-
-module Onkcop
-  # Your code goes here...
-end