+++ lib/rubocop/codetakt.rb
@@ -1,0 +1,4 @@
+module RuboCop
+  module Codetakt
+  end
+end