+++ lib/rubocop_codetakt.rb
@@ -1,0 +1,2 @@
+require "rubocop/codetakt/cli"
+require "rubocop/codetakt/version"